Refactor rebranding (#12)
* rename source folder from xops to mvxops

* rename xops mentions by mvxops

* add data folder initialisation in tests
