+++ mvxops/config/__init__.py
@@ -1,0 +1,5 @@
+"""
+author: Etienne Wallet
+
+This sub-package contains modules to handle the config logic for MvxOps
+"""
