+++ mvxops/resources/parser_help.txt
@@ -1,0 +1,18 @@
+## Command Line Interface for MvxOps ##
+
+mvxops [COMMAND] [SUBCOMMAND] [OPTIONS]
+
+Available commands:
+
+    config   -> Interact with the configuration
+                type 'mvxops config --help' for more information
+
+    data     -> Display data that have been saved localy.
+                type 'mvxops data --help' for more information
+
+    execute  -> Execute smart-contract interactions
+                type 'mvxops execute --help' for more information
+
+    version  -> Display the version of the mvxops package
+
+
