+++ mvxops/utils/__init__.py
@@ -1,0 +1,5 @@
+"""
+author: Etienne Wallet
+
+This sub-package is contains modules that serve as supports of the main functionalities of MvxOps
+"""
