--- xops/config/__init__.py
@@ -1,5 +1,0 @@
-"""
-author: Etienne Wallet
-
-This sub-package contains modules to handle the config logic for xOps
-"""
