--- xops/resources/parser_help.txt
@@ -1,18 +1,0 @@
-## Command Line Interface for xOps ##
-
-xops [COMMAND] [SUBCOMMAND] [OPTIONS]
-
-Available commands:
-
-    config   -> Interact with the configuration
-                type 'xops config --help' for more information
-
-    data     -> Display data that have been saved localy.
-                type 'xops data --help' for more information
-
-    execute  -> Execute smart-contract interactions
-                type 'xops execute --help' for more information
-
-    version  -> Display the version of the xops package
-
-
