--- xops/utils/__init__.py
@@ -1,5 +1,0 @@
-"""
-author: Etienne Wallet
-
-This sub-package is contains modules that serve as supports of the main functionalities of xOps
-"""
